requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Flask-Login==0.6.3
mysql-connector-python==9.0.0
cryptography==43.0.1
requests==2.32.3
Werkzeug==3.0.3
